Selemon branch

+++ include/gameObject/GoalTop.h
@@ -6,11 +6,12 @@
 class GoalTop : public Goal
 {
 public:
-	GoalTop(int x, int y, bool needScale);
+	GoalTop(float x, float y, bool needScale);
 	~GoalTop();
 
 	virtual void draw(sf::RenderWindow& window)const override;
 	virtual sf::Sprite& getSprite()override;
+
 private:
 
 	sf::Sprite m_sprite;
