+++ src/gameObject/GoalTop.cpp
@@ -1,6 +1,7 @@
 #include "gameObject/GoalTop.h"
-
-GoalTop::GoalTop(int x, int y, bool needScale)
+//----------------------------------------------------------------------------
+// Constructor initializes the goal top object
+GoalTop::GoalTop(float x, float y, bool needScale)
 {
 	m_sprite.setTexture(Resources::getInstance().getGoalTexture(2));
 	m_sprite.setPosition(x, y);
@@ -8,29 +9,32 @@ GoalTop::GoalTop(int x, int y, bool needScale)
 
 	if (needScale) {
 
-		m_sprite.scale(-1, 1);
+		m_sprite.scale(-1.f, 1.f);
 	}
 }
 
-
-
+//----------------------------------------------------------------------------
+// Static registration of LeftGoalTop in StaticFactory
 bool GoalTop::m_registeritLeftGoalTop = StaticFactory::registeritStatic("LeftGoalTop",
-	[]() -> std::shared_ptr<StaticObject> { return std::make_shared<GoalTop>(40, 580, false); });
+	[]() -> std::shared_ptr<StaticObject> { return std::make_shared<GoalTop>(40.f, 580.f, false); });
 
 
 bool GoalTop::m_registeritRightGoalTop = StaticFactory::registeritStatic("RightGoalTop",
-	[]() -> std::shared_ptr<StaticObject> { return std::make_shared<GoalTop>(1750, 580, true); });
+	[]() -> std::shared_ptr<StaticObject> { return std::make_shared<GoalTop>(1750.f, 580.f, true); });
 
+//----------------------------------------------------------------------------
+// Draw function to render the goal top object
 void GoalTop::draw(sf::RenderWindow& window)const {
 
 	window.draw(m_sprite);
 }
-
+//----------------------------------------------------------------------------
+// Getter function to retrieve the sprite of the goal top object
 sf::Sprite& GoalTop::getSprite() {
 
 	return m_sprite;
 }
-
+//----------------------------------------------------------------------------
 GoalTop::~GoalTop()
 {
 }
